add phast rollup service and move phast report logic
