+++ src/app/report-rollup/phast-rollup/phast-rollup-energy-table/phast-rollup-energy-table.component.ts
@@ -7,6 +7,7 @@ import { PhastResults, PHAST } from '../../../shared/models/phast/phast';
 import * as _ from 'lodash';
 import { ConvertUnitsService } from '../../../shared/convert-units/convert-units.service';
 import { PhastResultsData } from '../../report-rollup-models';
+import { PhastReportRollupService } from '../../phast-report-rollup.service';
 @Component({
   selector: 'app-phast-rollup-energy-table',
   templateUrl: './phast-rollup-energy-table.component.html',
@@ -25,12 +26,13 @@ export class PhastRollupEnergyTableComponent implements OnInit {
   energyCostUnit: string;
   energyPerMassUnit: string;
   energyPerTimeUnit: string;
-  constructor(private convertUnitsService: ConvertUnitsService, private reportRollupService: ReportRollupService, private suiteDbService: SuiteDbService) { }
+  constructor(private convertUnitsService: ConvertUnitsService, private reportRollupService: ReportRollupService, private suiteDbService: SuiteDbService,
+    private phastReportRollupService: PhastReportRollupService) { }
 
   ngOnInit() {
     this.settings = this.reportRollupService.checkSettings(this.settings);
     this.setUnits();
-    let phastResults: Array<PhastResultsData> = this.reportRollupService.phastResults.getValue();
+    let phastResults: Array<PhastResultsData> = this.phastReportRollupService.phastResults.getValue();
     let fuelResults: Array<PhastResultsData> = phastResults.filter(resultItem => { return resultItem.settings.energySourceType == 'Fuel' });
     this.setFuelSummary(fuelResults);
     let electricityResults: Array<PhastResultsData> = phastResults.filter(resultItem => { return resultItem.settings.energySourceType == 'Electricity' });
