+++ src/app/report-rollup/report-rollup.service.ts
@@ -26,27 +26,23 @@ import { SsmtService } from '../ssmt/ssmt.service';
 import { OpportunityCardsService, OpportunityCardData } from '../treasure-hunt/treasure-chest/opportunity-cards/opportunity-cards.service';
 import { OpportunityPaybackService } from '../treasure-hunt/treasure-hunt-report/opportunity-payback.service';
 import { PsatReportRollupService } from './psat-report-rollup.service';
+import { PhastReportRollupService } from './phast-report-rollup.service';
 
 
 @Injectable()
 export class ReportRollupService {
 
   reportAssessments: BehaviorSubject<Array<ReportItem>>;
-  phastAssessments: BehaviorSubject<Array<ReportItem>>;
   fsatAssessments: BehaviorSubject<Array<ReportItem>>;
   ssmtAssessments: BehaviorSubject<Array<ReportItem>>;
   treasureHuntAssessments: BehaviorSubject<Array<ReportItem>>;
 
   assessmentsArray: Array<ReportItem>;
-  phastArray: Array<ReportItem>;
   fsatArray: Array<ReportItem>;
   ssmtArray: Array<ReportItem>;
   treasureHuntArray: Array<ReportItem>;
 
 
-  selectedPhasts: BehaviorSubject<Array<PhastCompare>>;
-  phastResults: BehaviorSubject<Array<PhastResultsData>>;
-  allPhastResults: BehaviorSubject<Array<AllPhastResultsData>>;
 
   selectedFsats: BehaviorSubject<Array<FsatCompare>>;
   fsatResults: BehaviorSubject<Array<FsatResultsData>>;
@@ -60,15 +56,12 @@ export class ReportRollupService {
 
   calcsArray: Array<Calculator>;
   selectedCalcs: BehaviorSubject<Array<Calculator>>;
-  numPhasts: number = 0;
   numFsats: number = 0;
   numSsmt: number = 0;
   numTreasureHunt: number = 0;
   showSummaryModal: BehaviorSubject<string>;
   constructor(
-    private executiveSummaryService: ExecutiveSummaryService,
     private settingsService: SettingsService,
-    private phastResultsService: PhastResultsService,
     private directoryDbService: DirectoryDbService,
     private assessmentDbService: AssessmentDbService,
     private settingsDbService: SettingsDbService,
@@ -79,7 +72,8 @@ export class ReportRollupService {
     private opportunitySummaryService: OpportunitySummaryService,
     private opportunityCardsService: OpportunityCardsService,
     private opportunityPaybackService: OpportunityPaybackService,
-    private psatReportRollupService: PsatReportRollupService
+    private psatReportRollupService: PsatReportRollupService,
+    private phastReportRollupService: PhastReportRollupService
 
   ) {
     this.initSummary();
@@ -87,17 +81,14 @@ export class ReportRollupService {
 
   initSummary() {
     this.psatReportRollupService.initSummary();
+    this.phastReportRollupService.initSummary();
 
     this.reportAssessments = new BehaviorSubject<Array<ReportItem>>(new Array<ReportItem>());
-    this.phastAssessments = new BehaviorSubject<Array<ReportItem>>(new Array<ReportItem>());
     this.fsatAssessments = new BehaviorSubject<Array<ReportItem>>(new Array<ReportItem>());
     this.ssmtAssessments = new BehaviorSubject<Array<ReportItem>>(new Array<ReportItem>());
     this.treasureHuntAssessments = new BehaviorSubject<Array<ReportItem>>(Array<ReportItem>());
 
 
-    this.selectedPhasts = new BehaviorSubject<Array<PhastCompare>>(new Array<PhastCompare>());
-    this.phastResults = new BehaviorSubject<Array<PhastResultsData>>(new Array<PhastResultsData>());
-    this.allPhastResults = new BehaviorSubject<Array<AllPhastResultsData>>(new Array<AllPhastResultsData>());
 
     this.selectedFsats = new BehaviorSubject<Array<FsatCompare>>(new Array<FsatCompare>());
     this.fsatResults = new BehaviorSubject<Array<FsatResultsData>>(new Array<FsatResultsData>());
@@ -119,12 +110,7 @@ export class ReportRollupService {
     let settings = this.settingsDbService.getByAssessmentId(assessment);
     let tmpSettings = this.checkSettings(settings);
     this.assessmentsArray.push({ assessment: assessment, settings: tmpSettings });
-    // if (assessment.psat) {
-    //   this.psatArray.push({ assessment: assessment, settings: tmpSettings });
-    // } 
-    if (assessment.phast) {
-      this.phastArray.push({ assessment: assessment, settings: tmpSettings });
-    } else if (assessment.fsat) {
+    if (assessment.fsat) {
       this.fsatArray.push({ assessment: assessment, settings: tmpSettings });
     } else if (assessment.ssmt) {
       this.ssmtArray.push({
@@ -143,7 +129,7 @@ export class ReportRollupService {
 
   getReportData(directory: Directory) {
     this.assessmentsArray = new Array<ReportItem>();
-    this.phastArray = new Array<ReportItem>();
+    // this.phastArray = new Array<ReportItem>();
     // this.psatArray = new Array<ReportItem>();
     this.fsatArray = new Array<ReportItem>();
     this.ssmtArray = new Array<ReportItem>();
@@ -167,7 +153,8 @@ export class ReportRollupService {
         this.getChildDirectories(subDir);
       }
     });
-    this.phastAssessments.next(this.phastArray);
+    let phastArray: Array<ReportItem> = _.filter(this.assessmentsArray, (assessmentItem) => { return assessmentItem.assessment.phast != undefined });
+    this.phastReportRollupService.phastAssessments.next(phastArray);
     let psatArray: Array<ReportItem> = _.filter(this.assessmentsArray, (assessmentItem) => { return assessmentItem.assessment.psat != undefined });
     this.psatReportRollupService.psatAssessments.next(psatArray);
     this.fsatAssessments.next(this.fsatArray);
@@ -206,92 +193,6 @@ export class ReportRollupService {
 
 
 
-  //USED FOR PHAST SUMMARY
-  initPhastCompare(resultsArr: Array<AllPhastResultsData>) {
-    let tmpResults: Array<PhastCompare> = new Array<PhastCompare>();
-    resultsArr.forEach(result => {
-      let minCost = _.minBy(result.modificationResults, (result) => { return result.annualCost; });
-      if (minCost) {
-        let modIndex = _.findIndex(result.modificationResults, { annualCost: minCost.annualCost });
-        if (modIndex !== -1) {
-          let phastAssessments: Array<ReportItem> = this.phastAssessments.value;
-          let assessmentIndex = _.findIndex(phastAssessments, (val) => { return val.assessment.id === result.assessmentId; });
-          let item = phastAssessments[assessmentIndex];
-          if (result.isBaseline) {
-            tmpResults.push({ baseline: item.assessment.phast, modification: item.assessment.phast, assessmentId: result.assessmentId, selectedIndex: -1, name: item.assessment.name, assessment: item.assessment, settings: item.settings });
-          } else {
-            tmpResults.push({ baseline: item.assessment.phast, modification: item.assessment.phast.modifications[modIndex].phast, assessmentId: result.assessmentId, selectedIndex: modIndex, name: item.assessment.name, assessment: item.assessment, settings: item.settings });
-          }
-        }
-      }
-    });
-    this.selectedPhasts.next(tmpResults);
-  }
-
-  updateSelectedPhasts(item: ReportItem, modIndex: number) {
-    let tmpSelected = this.selectedPhasts.value;
-    if (modIndex !== -1) {
-      let selectedIndex = _.findIndex(tmpSelected, { assessmentId: item.assessment.id });
-      tmpSelected.splice(selectedIndex, 1, { baseline: item.assessment.phast, modification: item.assessment.phast.modifications[modIndex].phast, assessmentId: item.assessment.id, selectedIndex: modIndex, name: item.assessment.name, assessment: item.assessment, settings: item.settings });
-    }
-    else {
-      let selectedIndex = _.findIndex(tmpSelected, { assessmentId: item.assessment.id });
-      tmpSelected.splice(selectedIndex, 1, { baseline: item.assessment.phast, modification: item.assessment.phast, assessmentId: item.assessment.id, selectedIndex: modIndex, name: item.assessment.name, assessment: item.assessment, settings: item.settings });
-    }
-    this.selectedPhasts.next(tmpSelected);
-  }
-
-  initPhastResultsArr(phastArray: Array<ReportItem>) {
-    let tmpResultsArr = new Array<AllPhastResultsData>();
-    phastArray.forEach(val => {
-      if (val.assessment.phast.setupDone) {
-        let baselineResults = this.executiveSummaryService.getSummary(val.assessment.phast, false, val.settings, val.assessment.phast);
-        if (val.assessment.phast.modifications) {
-          if (val.assessment.phast.modifications.length !== 0) {
-            let modResultsArr = new Array<ExecutiveSummary>();
-            val.assessment.phast.modifications.forEach(mod => {
-              let tmpResults = this.executiveSummaryService.getSummary(mod.phast, true, val.settings, val.assessment.phast, baselineResults);
-              modResultsArr.push(tmpResults);
-            });
-            tmpResultsArr.push({ baselineResults: baselineResults, modificationResults: modResultsArr, assessmentId: val.assessment.id });
-          } else {
-            let modResultsArr = new Array<ExecutiveSummary>();
-            let tmpResults = this.executiveSummaryService.getSummary(val.assessment.phast, true, val.settings, val.assessment.phast, baselineResults);
-            modResultsArr.push(tmpResults);
-            tmpResultsArr.push({ baselineResults: baselineResults, modificationResults: modResultsArr, assessmentId: val.assessment.id, isBaseline: true });
-          }
-        } else {
-          let modResultsArr = new Array<ExecutiveSummary>();
-          let tmpResults = this.executiveSummaryService.getSummary(val.assessment.phast, true, val.settings, val.assessment.phast, baselineResults);
-          modResultsArr.push(tmpResults);
-          tmpResultsArr.push({ baselineResults: baselineResults, modificationResults: modResultsArr, assessmentId: val.assessment.id, isBaseline: true });
-        }
-      }
-    });
-    this.allPhastResults.next(tmpResultsArr);
-  }
-
-  getPhastResultsFromSelected(selectedPhasts: Array<PhastCompare>) {
-    let tmpResultsArr = new Array<PhastResultsData>();
-    selectedPhasts.forEach(val => {
-      let baselineResults = this.executiveSummaryService.getSummary(val.baseline, false, val.settings, val.baseline);
-      let modificationResults = this.executiveSummaryService.getSummary(val.modification, true, val.settings, val.baseline, baselineResults);
-      let baselineResultData = this.phastResultsService.getResults(val.baseline, val.settings);
-      let modificationResultData = this.phastResultsService.getResults(val.modification, val.settings);
-      tmpResultsArr.push({
-        baselineResults: baselineResults,
-        modificationResults: modificationResults,
-        assessmentId: val.assessmentId,
-        settings: val.settings,
-        name: val.name,
-        modName: val.modification.name,
-        assessment: val.assessment,
-        baselineResultData: baselineResultData,
-        modificationResultData: modificationResultData
-      });
-    });
-    this.phastResults.next(tmpResultsArr);
-  }
 
   //USED FOR FSAT SUMMARY
   initFsatCompare(resultsArr: Array<AllFsatResultsData>) {
