+++ src/app/report-rollup/report-summary/phast-summary/phast-summary.component.html
@@ -2,7 +2,7 @@
     <div class="card-header">
         <h3>Furnaces</h3>
         <span>
-            <span>{{reportRollupService.numPhasts}}</span>
+            <span>{{numPhasts}}</span>
         </span>
     </div>
     <div class="card-body">
@@ -10,7 +10,7 @@
             <div class="col-lg-6 col-md-8 col-sm-12">
                 <dl class="summary-list mx-auto">
                     <dt>Maximum Annual
-                        <br/>Cost Savings</dt>
+                        <br />Cost Savings</dt>
                     <dd>{{furnaceSavingsPotential| currency:'USD':'symbol':'2.0-0'}}</dd>
                     <dt>Annual Cost</dt>
                     <dd>{{totalCost | currency:'USD':'symbol':'2.0-0'}}</dd>
@@ -19,7 +19,7 @@
             <div class="col-lg-6 col-md-8 col-sm-12">
                 <dl class="summary-list mx-auto">
                     <dt>Maximum Annual
-                        <br/>Energy Savings</dt>
+                        <br />Energy Savings</dt>
                     <dd>{{energySavingsPotential| number:'2.0-0'}} {{settings.phastRollupUnit}}</dd>
                     <dt>Annual Energy</dt>
                     <dd>{{totalEnergy| number:'2.0-0'}} {{settings.phastRollupUnit}}
