+++ src/app/report-rollup/report-summary/report-summary.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 import { Subscription } from 'rxjs';
 import { Settings } from '../../shared/models/settings';
+import { PhastReportRollupService } from '../phast-report-rollup.service';
 import { PsatReportRollupService } from '../psat-report-rollup.service';
-import { PhastResultsData } from '../report-rollup-models';
 import { ReportRollupService } from '../report-rollup.service';
 
 @Component({
@@ -18,17 +18,25 @@ export class ReportSummaryComponent implements OnInit {
   showSummary: string = 'open';
 
   showPsatSummary: boolean;
+  showPhastSummary: boolean;
   psatAssessmentsSub: Subscription;
-  constructor(public reportRollupService: ReportRollupService, private psatReportRollupService: PsatReportRollupService) { }
+  phastAssessmentsSub: Subscription;
+  constructor(public reportRollupService: ReportRollupService, private psatReportRollupService: PsatReportRollupService,
+    private phastReportRollupService: PhastReportRollupService) { }
 
   ngOnInit() {
     this.psatAssessmentsSub = this.psatReportRollupService.psatAssessments.subscribe(val => {
       this.showPsatSummary = val.length != 0;
     });
+
+    this.phastAssessmentsSub = this.phastReportRollupService.phastAssessments.subscribe(val => {
+      this.showPhastSummary = val.length != 0;
+    });
   }
 
   ngOnDestroy(){
     this.psatAssessmentsSub.unsubscribe();
+    this.phastAssessmentsSub.unsubscribe();
   }
 
   showAssessmentModal(assessmentModalType: string){
